Cambiar tipo de dato de estado de String a Integer en entidades de admisiones, estructura institucional y módulo de estados
- Cambiar campo estado de String a Integer estadoId en InscripcionEntity, MatriculaEntity y EstudianteGrupoEntity
- Actualizar modelos de dominio Inscripcion, Matricula y EstudianteGrupo para usar Integer estadoId
- Modificar DTOs ActualizarEstadoInscripcionDto y ActualizarMatriculaDto para usar Integer estadoId
- Actualizar repositorios para buscar por estadoId en lugar de

+++ src/main/java/com/kleverkids/formacion_academica/modules/estados/application/input/consulta/ConsultaEspecializadaUseCase.java
@@ -6,7 +6,7 @@ import com.kleverkids.formacion_academica.modules.estados.infrastructure.outboun
 import java.util.List;
 
 public interface ConsultaEspecializadaUseCase {
-    List<EstadoDto> listarEntidadesConEstado(Long estadoId);
+    List<EstadoDto> listarEntidadesConEstado(Integer estadoId);
     List<EstadoDto> obtenerWorkflowCompleto(Long idModulo);
     List<EntidadEstadoEntity> listarEntidadesPorModulo(String entidadTipo, Long idModulo);
     boolean tieneEstadosConfigurados(Long idModulo);
